CODE RUB: 1 col CSV and hook refactor
