+++ ISL.ReIdentification.WebApps.Portals/ISL.ReIdentification.Portals.Client/src/components/csvReIdentification/csvReIdentificationDetailCardView.tsx
@@ -2,21 +2,21 @@ import React, { FunctionComponent, useState } from "react";
 import { Form, Button, Card, Spinner, Alert, OverlayTrigger, Tooltip, Modal } from "react-bootstrap";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faCircleInfo } from "@fortawesome/free-solid-svg-icons";
-import { reIdentificationService } from "../../services/foundations/reIdentificationService";
-import { AccessRequest } from "../../models/accessRequest/accessRequest";
 import { useMsal } from "@azure/msal-react";
 import UserAccessSearch from "../userAccessSearch/userAccessSearch";
 import { UserAccessView } from "../../models/views/components/userAccess/userAccessView";
 import { OverlayInjectedProps } from "react-bootstrap/esm/Overlay";
+import { useFileUpload } from "../../hooks/useFileUpload";
+import { useCsvData } from "../../hooks/useCsvData";
+import { reIdentificationService } from "../../services/foundations/reIdentificationService";
+import { AccessRequest } from "../../models/accessRequest/accessRequest";
 
 const CsvReIdentificationDetailCardView: FunctionComponent = () => {
-
     const [headerColumns, setHeaderColumns] = useState<string[]>([]);
     const [csvData, setCsvData] = useState<string | null>(null);
     const [selectedHeaderColumn, setSelectedHeaderColumn] = useState<string>("");
     const [selectedHeaderColumnIndex, setSelectedHeaderColumnIndex] = useState<number>(0);
     const [selectedUser, setSelectedUser] = useState<UserAccessView | undefined>();
-    const { submit, loading } = reIdentificationService.useRequestReIdentificationCsv();
     const [error, setError] = useState<string[]>([]);
     const [success, setSuccess] = useState("");
     const [savedSuccessfull, setSavedSuccessfull] = useState(false);
@@ -25,6 +25,11 @@ const CsvReIdentificationDetailCardView: FunctionComponent = () => {
     const [showHelpModal, setShowHelpModal] = useState<boolean>(false);
     const [reason, setReason] = useState<string>("");
     const account = useMsal();
+
+    const { handleFileChange } = useFileUpload(setError, setFileName, setCsvData, setHeaderColumns);
+    const { handleHeaderColumnChange } = useCsvData(csvData, headerColumns, setError, setSuccess);
+
+    const { submit, loading } = reIdentificationService.useRequestReIdentificationCsv();
 
     const handleSubmit = (e: React.FormEvent<HTMLFormElement>) => {
         e.preventDefault();
@@ -76,71 +81,6 @@ const CsvReIdentificationDetailCardView: FunctionComponent = () => {
         });
     };
 
-    const handleFileChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-        setError([]);
-        const file = e.target.files?.[0];
-        if (file) {
-            const sizeInBytes = file.size;
-            const sizeInMB = (sizeInBytes / (1024 * 1024)).toFixed(2);
-            console.log(`File size: ${sizeInBytes} bytes (${sizeInMB} MB)`);
-
-            if (sizeInBytes > 1024 * 1024) {
-                setError(["File size exceeds 1MB. Please upload a smaller file."]);
-                return;
-            }
-
-            if (file.name.endsWith(".csv")) {
-                setFileName(file.name);
-                const reader = new FileReader();
-                reader.onload = (event) => {
-                    let text = event.target?.result as string;
-                    text = text.replace(/\r/g, "");
-                    const rows = text.split("\n");
-                    const headers = rows[0].split(",");
-
-                    if (headers.length <= 1) {
-                        setError(["The CSV file does not contain a valid header row."]);
-                        return;
-                    }
-
-                    setHeaderColumns(headers);
-                    const uint8Array = new TextEncoder().encode(text);
-                    const base64String = btoa(String.fromCharCode(...uint8Array));
-                    setCsvData(base64String);
-                };
-                reader.readAsText(file);
-            } else {
-                setError(["Please upload a valid .csv file."]);
-            }
-        } else {
-            setError(["No file selected. Please upload a .csv file."]);
-        }
-    };
-
-    const handleHeaderColumnChange = (e: React.ChangeEvent<HTMLSelectElement>) => {
-        const selectedColumn = e.target.value;
-        setSelectedHeaderColumn(selectedColumn);
-        const index = headerColumns.indexOf(selectedColumn);
-        setSelectedHeaderColumnIndex(index);
-
-        if (csvData) {
-            const decodedData = atob(csvData);
-            const rows = decodedData.split("\n");
-
-            if (rows.length > 1) {
-                const nextRow = rows[1].split(",");
-                const value = nextRow[index];
-
-                if (/^\d{10}$/.test(value)) {
-                    setError([]);
-                    setSuccess(`The value "${value}" in the next row at the selected column index is 10 digits long and is a valid Pseudo Identifier.`);
-                } else {
-                    setSuccess("");
-                    setError([`The value "${value}" in the next row for the selected column index is not a valid Pseudo Identifier, please follow the guidance in the help section.`]);
-                }
-            }
-        }
-    };
     const handleCheckboxChange = (e: React.ChangeEvent<HTMLInputElement>) => {
         setHasHeaderRecord(e.target.checked);
     };
@@ -155,9 +95,6 @@ const CsvReIdentificationDetailCardView: FunctionComponent = () => {
         <>
             {!savedSuccessfull ? (
                 <>
-
-
-
                     <Card.Header>
                         <Card.Title className="text-start">
                             <OverlayTrigger placement="right" overlay={renderTooltip}>
@@ -167,7 +104,6 @@ const CsvReIdentificationDetailCardView: FunctionComponent = () => {
                     </Card.Header>
 
                     <Card.Body>
-
                         <Card.Subtitle className="text-start text-muted mb-3">
                             <small>
                                 Please upload your CSV file below, specify the reason for identifying these patients
@@ -196,7 +132,6 @@ const CsvReIdentificationDetailCardView: FunctionComponent = () => {
                                         placeholder="Upload CSV"
                                         accept=".csv"
                                         required />
-
                                 </div>
                                 <Form.Text className="text-muted">
                                     Please upload your CSV (other file types will be rejected).
@@ -209,7 +144,11 @@ const CsvReIdentificationDetailCardView: FunctionComponent = () => {
                                         <Form.Label><strong>Select Identifier Column From Csv:</strong></Form.Label>
                                         <Form.Select
                                             value={selectedHeaderColumn}
-                                            onChange={handleHeaderColumnChange}
+                                            onChange={(e) => {
+                                                setSelectedHeaderColumn(e.target.value);
+                                                setSelectedHeaderColumnIndex(headerColumns.indexOf(e.target.value));
+                                                handleHeaderColumnChange(e);
+                                            }}
                                             required>
                                             <option value="" disabled>Please select a column...</option>
                                             {headerColumns.map((column, index) => (
@@ -227,7 +166,6 @@ const CsvReIdentificationDetailCardView: FunctionComponent = () => {
                             )}
                             <Form.Group className="text-start">
                                 <Form.Label><strong>Re-identification Reason:</strong></Form.Label>
-
                                 <Form.Control
                                     as="textarea"
                                     value={reason}
@@ -236,7 +174,6 @@ const CsvReIdentificationDetailCardView: FunctionComponent = () => {
                                     rows={3}
                                     required
                                 />
-
                                 <Form.Text className="text-muted">
                                     Please supply a reason why you are requesting to re-identify, this will be visible in the email to the recipient.
                                 </Form.Text>
@@ -272,7 +209,6 @@ const CsvReIdentificationDetailCardView: FunctionComponent = () => {
                         <p>Alternatively, the recipient can launch their worklist in the portal and re-identify from there.</p>
                         <p><strong>Note:</strong> Files will last for 7 days and then be removed from the system.</p>
                     </Alert>
-
                 </>
             )}
 
@@ -307,10 +243,6 @@ const CsvReIdentificationDetailCardView: FunctionComponent = () => {
             </Modal>
         </>
     );
-
-    return <>
-        {error ? <Alert variant="danger">{error}</Alert> : <Alert variant="danger">Something went wrong.</Alert>}
-    </>;
-}
+};
 
 export default CsvReIdentificationDetailCardView;
